claude_agent_course/module_03_custom_tools/projects/project_04_calculator_plus/pyproject.toml
[project]
name = "calculator-plus"
version = "0.1.0"
description = "高级计算器 - 自定义工具实战"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
